# Improve L33tMatcher performance a bit