I tried to improve the performance L33tMatcher.
I ran the unit tests five times each before and after this change and compared the execution times.

Before:
1. 8s 310ms
2. 8s 229ms
3. 8s 189ms
4. 8s 240ms
5. 8s 350ms

After
1. 7s787ms
2. 7s768ms
3. 7s891ms
4. 7s838ms
5. 7s878ms